improved delete directory function

+++ src/raspui.cpp
@@ -1,6 +1,7 @@
 #include <QDebug>
 #include <QDir>
 #include <QDirIterator>
+#include <QMessageBox>
 #include "raspui.h"
 #include "ui_mainwindow.h"
 
@@ -12,11 +13,15 @@ MainWindow::MainWindow(QWidget *parent) :
 {
     ui->setupUi(this);
     ui->Path->setText(currentDir.absolutePath());
+    ui->Log->setSelectionMode(QAbstractItemView::ExtendedSelection);
     connect(ui->Path,SIGNAL(editingFinished()),this,SLOT(textChangedSlot()));
+
     //changing context menu policy and connecting signal-slot
     ui->Log->setContextMenuPolicy(Qt::CustomContextMenu);
     connect(ui->Log, SIGNAL(customContextMenuRequested(QPoint)),this,SLOT(showContextMenu(QPoint)));
+
     connect(ui->Log, SIGNAL(itemDoubleClicked(QListWidgetItem*)), this, SLOT(onItemDoubleClicked(QListWidgetItem*)));
+
 
 }
 
@@ -26,7 +31,6 @@ MainWindow::~MainWindow(){
 
 void MainWindow::on_Execute_clicked(){
     ui->Log->clear();
-    ui->Path->setText(currentDir.absolutePath());
     QDirIterator it(currentDir, QDirIterator::NoIteratorFlags);
     while (it.hasNext()) {
         ui->Log->addItem(it.next());
@@ -36,33 +40,37 @@ void MainWindow::on_Execute_clicked(){
 void MainWindow::on_Back_clicked(){
     currentDir.cdUp();
     on_Execute_clicked();
+    ui->Path->setText(currentDir.absolutePath());
 }
 
  void MainWindow::textChangedSlot(){
     currentDir.cd(ui->Path->text());
-    //on_Execute_clicked();
+    on_Execute_clicked();
  }
 
  void MainWindow::showContextMenu(const QPoint &pos){
      QPoint globalPos = ui->Log-> mapToGlobal(pos);
      QMenu dropMenu;
      dropMenu.addAction("Delete", this, SLOT(eraseItem()));
+     //dropMenu.addAction("Create", this, SLOT(createItem()));
      dropMenu.exec(globalPos);
  }
 
- void MainWindow::eraseItem(){
-     //TODO Handle error, progress bar, pop up window
-     for (int i=0; i < ui->Log->selectedItems().size(); i++){
-         QListWidgetItem *item = ui->Log->takeItem(ui->Log->currentRow());
-         QDir dirToDelete(item->text());
-         dirToDelete.removeRecursively();
-         delete item;
+ void MainWindow::eraseItem(){                  //TODO Handle error, progress bar
+     QMessageBox::StandardButton confirmation;
+     QList<QListWidgetItem*> items = ui->Log->selectedItems();
+     foreach(QListWidgetItem * item, items){
+        confirmation = QMessageBox::question(this, "RASP", "Are you sure you want to delete "+ item->text(), QMessageBox::Yes|QMessageBox::No);
+        if(confirmation == QMessageBox::Yes){
+            QDir dirToDelete(item->text());
+            dirToDelete.removeRecursively();
+            delete item;
+         }
      }
-
  }
 
  void MainWindow::onItemDoubleClicked(QListWidgetItem *item){
      currentDir.cd(item->text());
      on_Execute_clicked();
-
+     ui->Path->setText(currentDir.absolutePath());
  }
